+++ src/jobs/Recruit.Vacancies.Jobs/Infrastructure/IUpdateQueryStore.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using Esfa.Recruit.Vacancies.Jobs.Models;
 
-namespace Esfa.Recruit.Vacancies.Jobs
+namespace Esfa.Recruit.Vacancies.Jobs.Infrastructure
 {
     public interface IUpdateQueryStore
     {
