+++ src/jobs/Recruit.Vacancies.Jobs/TrainingTypes/ApprenticeshipApiCollectionExtensions.cs
